Simplify gettext setup.
It's better to set up gettext once at application start (that way, we
can have translations before the start() method is called).

However, this method requires our package to be installed in a
plain directory on the filesystem.  In theory, packages may also be run
from a .zip file, and we should use importlib.resources to access data
in our package.  Unfortunately, the standard Python gettext module is
not compatible with this approach, and there seems to be no standard
method to install the gettext *.mo files together with our package.