enca_plugin: Add trailing '_' to widget object names.
This clearly separates the end of our chosen object name in case Qt
adds a numbered suffix.

Otherwise, there was ambiguity for widget names such as 'l12' -> does
the name correspond to config key 'l12', or is is the third widget
named 'l1', for which Qt added a suffix '2'?  Adding an underscore
clearly distinguishes both cases:

 - the widget for config key 'l12' now has the name
   'l12_{possible_suffix}'

 - the third widget for config key 'l1' with a numerical suffix added
   by Qt now has the name 'l1_2'.